Moved image adapter out of activity.
