+++ app/src/main/java/com/codefororlando/petadoption/view/DetailsActivity.java
@@ -1,19 +1,14 @@
 package com.codefororlando.petadoption.view;
 
-import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
 import android.support.annotation.StringRes;
-import android.support.v4.view.PagerAdapter;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.ActionBar;
 import android.support.v7.widget.Toolbar;
-import android.view.LayoutInflater;
 import android.view.MenuItem;
 import android.view.View;
-import android.view.ViewGroup;
-import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 import android.widget.Toast;
@@ -26,9 +21,6 @@ import com.codefororlando.petadoption.data.model.Location;
 import com.codefororlando.petadoption.data.model.Shelter;
 import com.codefororlando.petadoption.presenter.details.DetailsPresenter;
 import com.codefororlando.petadoption.recyclerview.HorizontalViewPagerIndicator;
-import com.squareup.picasso.Picasso;
-
-import java.util.List;
 
 import nucleus.factory.RequiresPresenter;
 import nucleus.view.NucleusAppCompatActivity;
@@ -196,51 +188,6 @@ public class DetailsActivity extends NucleusAppCompatActivity<DetailsPresenter> 
         hideCallAction();
         hideWebAction();
         hideEmailAction();
-    }
-
-    private static class PetImageViewPagerAdapter extends PagerAdapter {
-
-        private final Context context;
-        private final List<String> images;
-        private final int placeholderImage;
-        private final LayoutInflater mLayoutInflater;
-
-        public PetImageViewPagerAdapter(Context context, List<String> images, int placeholderImage) {
-            mLayoutInflater = (LayoutInflater) context.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
-            this.context = context;
-            this.images = images;
-            this.placeholderImage = placeholderImage;
-        }
-
-        @Override
-        public int getCount() {
-            return images.size();
-        }
-
-        @Override
-        public boolean isViewFromObject(View view, Object object) {
-            return view == object;
-        }
-
-        @Override
-        public Object instantiateItem(ViewGroup container, int position) {
-            View itemView = mLayoutInflater.inflate(R.layout.image_pager_item, container, false);
-            ImageView imageView = (ImageView) itemView.findViewById(R.id.image_content);
-            Picasso.with(context)
-                    .load(images.get(position))
-                    .resize(1000, 1000)
-                    .onlyScaleDown()
-                    .centerCrop()
-                    .placeholder(placeholderImage)
-                    .into(imageView);
-            container.addView(itemView);
-            return itemView;
-        }
-
-        @Override
-        public void destroyItem(ViewGroup container, int position, Object object) {
-            container.removeView((View) object);
-        }
     }
 
 }
